+++ docs/source/Core.rst
@@ -69,6 +69,8 @@ Core
 
             - `TransferFunctions`
 
+            - `TimerFunctions`
+
             - `TransformFunctions`
 
       - `StatefulFunctions`
