+++ docs/source/TimerFunctions.rst
@@ -1,0 +1,10 @@
+TimerFunctions
+==============
+
+.. toctree::
+   :maxdepth: 3
+
+.. automodule:: psyneulink.core.components.functions.nonstateful.timerfunctions
+   :members: ExponentialRise, ExponentialDecay, LinearRise, LinearDecay, LogarithmicRise LogarithmicDecay
+   :private-members:
+   :exclude-members: Parameters
