+++ docs/source/TransferFunctions.rst
@@ -4,8 +4,8 @@ TransferFunctions
 .. toctree::
    :maxdepth: 3
 
-.. automodule:: psyneulink.core.components.functions.transferfunctions
-   :members: TransferFunction, Identity, Linear, Exponential, ExponentialDecay, AcceleratingDecay, Logistic, Tanh,ReLU, Angle, Gaussian, GaussianDistort, BinomialDistort, Dropout, SoftMax, LinearMatrix, TransferWithCosts,
+.. automodule:: psyneulink.core.components.functions.nonstateful.transferfunctions
+   :members: TransferFunction, Identity, Linear, Exponential, Logistic, Tanh,ReLU, Angle, Gaussian, GaussianDistort, BinomialDistort, Dropout, SoftMax, LinearMatrix, TransferWithCosts,
 CostFunctions
    :private-members:
    :exclude-members: Parameters
