+++ psyneulink/core/components/mechanisms/processing/transfermechanism.py
@@ -317,7 +317,7 @@ By default, `execute_until_finished <Component.execute_until_finished>` is True,
 <TranserMechanism.integrator_mode>` is set to True a TransferMechanism will execute until it terminates, using a
 `convergence criterion <TransferMechanism_Convergence_Termination>`.  However, the Mechanism's method of termination
 can be configured using its `termination_measure <TransferMechanism.termination_measure>` and `termination_comparison_op
-<TransferMechanism.termination_comparison_op>` `Parameters` can be used to congifure other termination conditions.
+<TransferMechanism.termination_comparison_op>` `Parameters` can be used to configure other termination conditions.
 There are two broad types of termination condition: convergence and boundary termination.
 
 .. _TransferMechanism_Convergence_Termination:
