+++ psyneulink/core/globals/keywords.py
@@ -24,9 +24,10 @@ https://princetonuniversity.github.io/PsyNeuLink/Keywords.html
 #    that are the ones actually used by the code.
 
 __all__ = [
-    'ACCELERATING_DECAY_FUNCTION', 'ACCUMULATOR_INTEGRATOR', 'ACCUMULATOR_INTEGRATOR_FUNCTION',
-    'ADAPTIVE', 'ADAPTIVE_INTEGRATOR_FUNCTION', 'ADAPTIVE_MECHANISM', 'ADD_INPUT_PORT', 'ADD_OUTPUT_PORT', 'ADDITIVE',
-    'ADDITIVE_PARAM', 'AFTER', 'ALL', 'ALLOCATION_SAMPLES', 'ALLOW_PROBES', 'ANGLE', 'ANGLE_FUNCTION','ANY',
+    'ACCUMULATOR_INTEGRATOR', 'ACCUMULATOR_INTEGRATOR_FUNCTION',
+    'ADAPTIVE', 'ADAPTIVE_INTEGRATOR_FUNCTION', 'ADAPTIVE_MECHANISM',
+    'ADD_INPUT_PORT', 'ADD_OUTPUT_PORT', 'ADDITIVE', 'ADDITIVE_PARAM',
+    'AFTER', 'ALL', 'ALLOCATION_SAMPLES', 'ALLOW_PROBES', 'ANGLE', 'ANGLE_FUNCTION','ANY',
     'ARG_MAX', 'ARG_MAX_ABS', 'ARG_MAX_INDICATOR', 'ARG_MAX_ABS_INDICATOR',
     'ARG_MIN', 'ARG_MIN_ABS', 'ARG_MIN_INDICATOR', 'ARG_MIN_ABS_INDICATOR',
     'ARGUMENT_THERAPY_FUNCTION', 'ARRANGEMENT', 'ASSERT', 'ASSIGN', 'ASSIGN_VALUE', 'AUTO','AUTO_ASSIGN_MATRIX',
@@ -79,8 +80,10 @@ __all__ = [
     'LEARNING', 'LEARNING_FUNCTION', 'LEARNING_FUNCTION_TYPE', 'LEARNING_OBJECTIVE', 'LEARNING_MECHANISM',
     'LEARNING_MECHANISMS', 'LEARNING_PATHWAY', 'LEARNING_PROJECTION', 'LEARNING_PROJECTION_PARAMS', 'LEARNING_RATE',
     'LEARNING_SCALE', 'LEARNING_SCALE_LITERALS', 'LEARNING_SCALE_NAMES', 'LEARNING_SIGNAL', 'LEARNING_SIGNAL_SPECS',
-    'LEARNING_SIGNALS', 'LESS_THAN', 'LESS_THAN_OR_EQUAL', 'LINEAR', 'LINEAR_COMBINATION_FUNCTION', 'LINEAR_FUNCTION',
-    'LOG_ENTRIES', 'LOGISTIC_FUNCTION', 'Loss', 'LOSSES', 'LOW', 'LVOC_CONTROL_MECHANISM',
+    'LEARNING_SIGNALS', 'LESS_THAN', 'LESS_THAN_OR_EQUAL',
+    'LINEAR', 'LINEAR_COMBINATION_FUNCTION', 'LINEAR_DECAY_FUNCTION', 'LINEAR_FUNCTION', 'LINEAR_RISE_FUNCTION',
+    'LOG_ENTRIES', 'LOGARITHMIC_DECAY_FUNCTION', 'LOGARITHMIC_RISE_FUNCTION', 'LOGISTIC_FUNCTION',
+    'Loss', 'LOSSES', 'LOW', 'LVOC_CONTROL_MECHANISM',
     'MAPPING_PROJECTION', 'MAPPING_PROJECTION_PARAMS', 'MASKED_MAPPING_PROJECTION',
     'MATRIX', 'MATRIX_KEYWORD_NAMES', 'MATRIX_KEYWORD_SET', 'MATRIX_KEYWORD_VALUES', 'MATRIX_KEYWORDS',
     'MatrixKeywords', 'MATRIX_TRANSFORM_FUNCTION',  'MATRIX_WEIGHTS',
@@ -129,7 +132,7 @@ __all__ = [
     'STANDARD_ARGS', 'STANDARD_DEVIATION', 'STANDARD_OUTPUT_PORTS', 'START', 'STORE', 'SUBTRACTION', 'SUM',
     'TARGET', 'TARGET_MECHANISM', 'TARGET_LABELS_DICT', 'TERMINAL', 'TARGETS',
     'TERMINATION_MEASURE', 'TERMINATION_THRESHOLD', 'TERMINATION_COMPARISION_OP', 'TERSE', 'TEXT', 'THRESHOLD',
-    'TIME', 'TIME_STEP_SIZE', 'TIME_STEPS_DIM', 'TOLERANCE', 'TRAINED_OUTPUTS', 'TRAINING_SET',
+    'TIME', 'TIME_STEP_SIZE', 'TIME_STEPS_DIM', 'TIMER_FUNCTION_TYPE', 'TOLERANCE', 'TRAINED_OUTPUTS', 'TRAINING_SET',
     'TRANSFER_FUNCTION_TYPE', 'TRANSFER_MECHANISM', 'TRANSFER_WITH_COSTS_FUNCTION', 'TRIAL', 'TRIALS_DIM',
     'UNCHANGED', 'UNIFORM_DIST_FUNCTION', 'UPDATE', 'USER_DEFINED_FUNCTION', 'USER_DEFINED_FUNCTION_TYPE',
     'VALUES', 'VALIDATE', 'VALIDATION', 'VALUE', 'VALUE_ASSIGNMENT', 'VALUE_FUNCTION', 'VARIABLE', 'VARIANCE',
@@ -713,6 +716,7 @@ STATEFUL_FUNCTION_TYPE = "STATEFUL FUNCTION TYPE"
 MEMORY_FUNCTION_TYPE = "MEMORY FUNCTION TYPE"
 INTEGRATOR_FUNCTION_TYPE = "INTEGRATOR FUNCTION TYPE"
 TRANSFER_FUNCTION_TYPE = "TRANSFER FUNCTION TYPE"
+TIMER_FUNCTION_TYPE = "TIMER FUNCTION TYPE"
 LEABRA_FUNCTION_TYPE = "LEABRA FUNCTION TYPE"
 DISTRIBUTION_FUNCTION_TYPE = "DISTRIBUTION FUNCTION TYPE"
 OBJECTIVE_FUNCTION_TYPE = "OBJECTIVE FUNCTION TYPE"
@@ -768,8 +772,8 @@ IDENTITY_FUNCTION = 'Identity Function'
 LINEAR_FUNCTION = "Linear Function"
 LEABRA_FUNCTION = "Leabra Function"
 EXPONENTIAL_FUNCTION = "Exponential Function"
-EXPONENTIAL_DECAY_FUNCTION = "ExponentialDecay Function"
-ACCELERATING_DECAY_FUNCTION = "AcceleratingDecay Function"
+LOGARITHMIC_FUNCTION = "Logarithmic Function"
+LOGARITHMIC_RISE_FUNCTION = "LogarithmicRise Function"
 LOGISTIC_FUNCTION = "Logistic Function"
 TANH_FUNCTION = "Tanh Function"
 RELU_FUNCTION = "ReLU Function"
@@ -780,6 +784,14 @@ BINOMIAL_DISTORT_FUNCTION = 'BinomialDistort Function'
 DROPOUT_FUNCTION = 'Dropout Function'
 SOFTMAX_FUNCTION = 'SoftMax Function'
 TRANSFER_WITH_COSTS_FUNCTION = "TransferWithCosts Function"
+
+# TimerFunctions:
+EXPONENTIAL_DECAY_FUNCTION = "ExponentialDecay Function"
+EXPONENTIAL_RISE_FUNCTION = "ExponentialRise Function"
+LINEAR_DECAY_FUNCTION = "LinearDecay Function"
+LINEAR_RISE_FUNCTION = "LinearRise Function"
+LOGARITHMIC_DECAY_FUNCTION = "LogarithmicDecay Function"
+LOGARITHMIC_RISE_FUNCTION = "LogarithmicRise Function"
 
 # SelectionFunctions:
 ONE_HOT_FUNCTION = "OneHot Function"
